hwp2docx/requirements.txt
python-docx>=1.1.0
pyhwp>=0.1b15
six>=1.16.0
